% Generated by roxygen2 (4.0.1): do not edit by hand
\name{importEmfitJSON}
\alias{importEmfitJSON}
\title{Function to import Emfit JSON data into R dataframe}
\usage{
importEmfitJSON(filename, timezone)
}
\arguments{
\item{filename}{String containing the path and filename of JSON data file.}

\item{timezone}{The timezone of recorded data.}
}
\description{
Function imports Emfit JSON data into R dataframe. The data frame contains columns for heart rate,
breathing rate, activity, timestamp and information about quality of each measured quantity.
}

